+++ model/src/main/java/es/iespuerto/ets/GameManager.java
@@ -5,18 +5,17 @@ package es.iespuerto.ets;
  * suspension y terminacion del videojuego
  * 
  * @author Jonathan
- * @param simpleGameManager referencia del game manager
  */
 public class GameManager {
-    public byte simpleGameManager;
+    public int simpleGameManager;
 
     /**
      * Constructor para establecer el GameManager se introduce un id en forma de
-     * byte
+     * int
      * 
      * @param simpleGameManager Propiedad de la clase
      */
-    public GameManager(byte simpleGameManager) {
+    public GameManager(int simpleGameManager) {
         this.simpleGameManager = simpleGameManager;
     }
 
@@ -48,11 +47,11 @@ public class GameManager {
 
     }
 
-    public byte getSimpleGameManager() {
+    public int getSimpleGameManager() {
         return simpleGameManager;
     }
 
-    public void setSimpleGameManager(byte simpleGameManager) {
+    public void setSimpleGameManager(int simpleGameManager) {
         this.simpleGameManager = simpleGameManager;
     }
